+++ src/views/evidence_metadata_dialog.c
@@ -13,6 +13,7 @@
 #include "views/evidence_identity_ocr_dialog.h"
 #include "views/person_vocabulary_adapter.h"
 #include "views/document_authenticity_editor.h"
+#include "views/document_identity_misuse_editor.h"
 #include "widgets/ocr_provenance_overlay.h"
 
 struct EvidenceMetadataDialogResult
@@ -50,6 +51,7 @@ typedef struct
     gpointer user_data;
     gboolean completed;
     DocumentAuthenticityEditor *authenticity_editor;
+    DocumentIdentityMisuseEditor *identity_misuse_editor;
 } EvidenceMetadataDialogContext;
 
 static void evidence_metadata_dialog_complete(
@@ -92,6 +94,7 @@ static void evidence_metadata_dialog_context_free(gpointer data)
     g_clear_pointer(&context->ocr_run_identifiers, g_ptr_array_unref);
     g_clear_object(&context->ocr_run_labels);
     document_authenticity_editor_free(context->authenticity_editor);
+    document_identity_misuse_editor_free(context->identity_misuse_editor);
     g_free(context->investigation_root);
     g_free(context->evidence_identifier);
     g_free(context);
@@ -553,6 +556,16 @@ gboolean evidence_metadata_dialog_present_with_ocr(
             gtk_scrolled_window_set_child(GTK_SCROLLED_WINDOW(auth_scroll),
                 document_authenticity_editor_get_widget(context->authenticity_editor));
             gtk_box_append(GTK_BOX(main_box), auth_scroll);
+        }
+        context->identity_misuse_editor = document_identity_misuse_editor_new(
+            database, evidence_record_get_identifier(record));
+        if (context->identity_misuse_editor != NULL) {
+            GtkWidget *misuse_scroll = gtk_scrolled_window_new();
+            gtk_widget_set_size_request(misuse_scroll, -1, 300);
+            gtk_scrolled_window_set_child(GTK_SCROLLED_WINDOW(misuse_scroll),
+                document_identity_misuse_editor_get_widget(
+                    context->identity_misuse_editor));
+            gtk_box_append(GTK_BOX(main_box), misuse_scroll);
         }
         IdentityTraceabilityDao *trace_dao = identity_traceability_dao_new(database);
         if (trace_dao != NULL) {